c.chatgpt: capture hashed user ID
OpenAI recommends including an anonymized user ID when calling their API. That way if somebody is doing something abusive, the perpetrators can be tracked down by the client. Since this bot is a public service, this is probably a good idea. We use a SHA256 of the discord user ID snowflake.

- Added "namehash" attribute to the GPTUser class
- Updated the __init__ method of GPTUser to calculate and assign the namehash using sha256 hashing algorithm
- Modified the send_to_chatgpt method in ChatGPT class to accept an additional "user" parameter
- Updated the send_to_chatgpt method to pass the user parameter when making API call for chat completion
- Modified the ChatGPT command methods (e.g., on_message, summarize_messages) to pass the user_id as an argument when calling send_to_chatgpt